Add a method to easily update the properties of a virtual instance
So far updating a virtual instance state, cpu count or memory amount,
forces deleting the record from the database and recreating it.

This commit adds a function that only updates those simple properties
without removing the record.

In order to be able to change the VM state, remove the update="false"
attribute from VirtualInstanceInfos state relation.