Add virtualization guest creation salt implementation
The salt implementation of the VM guest creation calls either virt.running
state or virt.update execution module, which have been both thought to
take the similar parameters.

What makes a VM update is the presence of an UUID in the action details.
The only thing that prevents us from using virt.running for the update
case is that VM update doesn't mean the VM is running after the change.