Add disks to GuestProperties
Allow user to edit the virtual disks of a VM. If at least one of the
disk configurations is unsupported, disable editing of all disks to
avoid damaging the unsupported ones.

So far only disks of 'file' type are supported. Even though the code is
ready for more in a number of places, salt can't handle other types ATM.