Add network to GuestProperties
Allow editing the existing network interfaces of a VM. If at least one
of the network interfaces is not supported, disable editing of all NICs
to avoid damaging the unsupported ones.

So far only the NICs of type 'network' are supported. We will try to
push users to define as much as possible into the virtual network
itself. The advanced NIC features like vlan ports or IP configuration
are not supported yet either.