Implement incremental package update cache update and truncating errata
Previously, when aligning packages, the update cache (rhnServerNeededCache)
was cleared for target channel. Now the cache is updated incrementally.

Aligning errata has also been implemented: when aliging a channel with
errata A to a channel with errata B, the target channel will now end up
with only errata B (A will be removed from it).